v1.4. Satnam's version that allows the outside lights to be switched on and off (after 15 secs) using the wall light, and switched on using the app.
---------

Hi Satnam,

Thanks for this. Looks like this is an improvement.

Using the switch - the outside lights switch on but you can't switch them off straight away. You have to wait 10-15 secs and then use the switch to switch them off.

Using the app - the outside lights switch on but you can't switch them off. You have to use the wall switch to switch them off after you have waited the 10-15 seconds.

I've uploaded this version to github with these notes.

Cheers,

Steve

----------

Hi Steve, could you please upload the attached to the controller. This is based on the original file with following changes:

Hot Water Pump – Time Change
Dimmer hardware readdressing

I’ve tried to replicate the same issue in the office without any success (as don’t have that particular dimmer module).

I’ve spoken to technical who of the opinion is has something to do with the obsolescent Dimmer module, but have advise to make a couple of non-standard changes to the hardware addressing to force it.

If working, should work from the app and switch again with feedback.