That's loads better. I can switch the outside lights on and off as I would expect using both the iOS app and the switch on the wall. The on/off status is shown in the app, but not on the wall switch LED. So all working except for the LED, which is not a major issue, that I'm happy to live with (don't think it's worth the additional cost to fix. Thanks for resolving.
Just one other question. Is there a reason the hot water pump timer isn't shown in the iOS app? It's not a major issue, just curious.

Thanks,
Steve

-----------

Morning Steve, please find attached another file to try. Effectively omitted the LED feedback from Dimmer as Technical Support said that might be the cause with.

Let us know how you get on.

Failing that we’re looking at £243.55 for a replacement dimmer and £200.00 for labour. Availability is from 11/02/2022. Not an insignificant amount to be fair, so you’ll need to decide whether you want it replaced or can live with it as it is?

Please advise.
Thanks
Satnam Singh